# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "本地化的 Bangumi 数据服务。利用 Archive 每周导出的 wiki 数据导入本地 SQLite，\n提供条目/人物/角色/章节等 REST 查询接口与单二进制管理接口。\n统一响应包装：成功 {\"ok\": true, \"data\": ...}；失败 {\"ok\": false, \"error\": \"...\"}。\n数据接口均为只读；管理接口（/api/admin/*）需要 X-Admin-Token 请求头（未配置 token 时放行）。\n交互式文档：/docs 或 /swagger/index.html；本文档由 swag 从代码注解自动生成（make docs）。",
        "title": "Bangumi 本地数据搜索 API",
        "contact": {
            "name": "GitHub",
            "url": "https://github.com/nz3u/bangumi-data"
        },
        "license": {
            "name": "AGPL-3.0",
            "url": "https://www.gnu.org/licenses/agpl-3.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/characters/search": {
            "get": {
                "description": "q 匹配原名与 infobox 简体中文名（≥3 字符走全文索引，更短回退 LIKE 子串匹配）。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色"
                ],
                "summary": "角色搜索",
                "parameters": [
                    {
                        "type": "string",
                        "description": "关键词",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "enum": [
                            1,
                            2,
                            3
                        ],
                        "type": "integer",
                        "description": "角色类型：1 角色 / 2 机体 / 3 组织",
                        "name": "role",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 30,
                        "description": "每页数量",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.apiEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.characterSearchData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.apiEnvelope"
                        }
                    }
                }
            }
        },
        "/api/characters/{id}": {
            "get": {
                "description": "含基本信息、infobox（结构化字段）、简介、出演作品、声优/演员（按人物去重）与角色关联（双向）。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "角色"
                ],
                "summary": "角色详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "角色 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.apiEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.characterDetail"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "无效的 id",
                        "schema": {
                            "$ref": "#/definitions/api.apiEnvelope"
                        }
                    },
                    "404": {
                        "description": "角色不存在",
                        "schema": {
                            "$ref": "#/definitions/api.apiEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.apiEnvelope"
                        }
                    }
                }
            }
        },
...
        "/api/episodes/search": {
            "get": {
                "description": "q 同时命中章节标题与所属条目标题（原名/中文名/别名，归一化口径同条目搜索）；字段含义同 Archive 的 episode 表：sort=集数、disc=所在光盘、airdate=播出时间（原文文本，范围匹配对 ISO 日期有效）、duration=时长。≥3 字符走 trigram 全文索引，短词自动走命中集路径。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "章节"
                ],
                "summary": "章节搜索与筛选",
                "parameters": [
                    {
                        "type": "string",
                        "description": "关键词（全文搜索）",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "所属条目 ID",
                        "name": "subject_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "作品类型（所属条目）：1 书籍 / 2 动画 / 3 音乐 / 4 游戏 / 6 三次元",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "enum": [
                            0,
                            1,
                            2,
                            3,
                            4,
                            5,
                            6
                        ],
                        "type": "integer",
                        "description": "章节类型：0 正篇 / 1 SP / 2 OP / 3 ED / 4 Trailer / 5 MAD / 6 其他",
                        "name": "ep_type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "光盘号",
                        "name": "disc",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "播出时间下界",
                        "name": "airdate_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "播出时间上界",
                        "name": "airdate_to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序：id / sort=集数 / airdate=播出日期 / subject=条目 / popularity=条目人气，省略=智能排序",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "asc",
                        "description": "排序方向",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 30,
                        "description": "每页数量",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.apiEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.episodeSearchData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.apiEnvelope"
                        }
                    }
                }
            }
        },
...
        "/api/persons/search": {
            "get": {
                "description": "q 匹配原名与 infobox 简体中文名（≥3 字符走全文索引，更短回退 LIKE 子串匹配）。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "人物"
                ],
                "summary": "人物搜索",
                "parameters": [
                    {
                        "type": "string",
                        "description": "关键词",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "enum": [
                            1,
                            2,
                            3
                        ],
                        "type": "integer",
                        "description": "人物类型：1 个人 / 2 公司 / 3 组合",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 30,
                        "description": "每页数量",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.apiEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.personSearchData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.apiEnvelope"
                        }
                    }
                }
            }
        },
        "/api/persons/{id}": {
            "get": {
                "description": "含基本信息、infobox（结构化字段）、简介与参与/出演统计。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "人物"
                ],
                "summary": "人物详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "人物 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.apiEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.personDetail"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "无效的 id",
                        "schema": {
                            "$ref": "#/definitions/api.apiEnvelope"
                        }
                    },
                    "404": {
                        "description": "人物不存在",
                        "schema": {
                            "$ref": "#/definitions/api.apiEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.apiEnvelope"
                        }
                    }
                }
            }
        },
        "/api/persons/{id}/collaboration": {
            "get": {
                "description": "人物简介 + 分页的合作人物及共同参与条目（含声优等出演，职位 id 已转常量文本）。合作人物按共同条目数降序。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "人物合作"
                ],
                "summary": "人物合作",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "人物 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 30,
                        "description": "每页数量",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "当前人物职位标签（key 形如 2:1 或 cv，逗号分隔多选，- 前缀为排除），先经 /positions 获取",
                        "name": "positions_a",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "合作人物职位标签，语法同 positions_a；两组之间取交集",
                        "name": "positions_b",
                        "in": "query"
                    },
                    {
                        "enum": [
                            1,
                            2
                        ],
                        "type": "integer",
                        "description": "合作人物类型：1 个人 / 2 公司（组织）",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.apiEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.collaborationData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "无效的 id",
                        "schema": {
                            "$ref": "#/definitions/api.apiEnvelope"
                        }
                    },
                    "404": {
                        "description": "人物不存在",
                        "schema": {
                            "$ref": "#/definitions/api.apiEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.apiEnvelope"
                        }
                    }
                }
            }
        },
        "/api/persons/{id}/collaboration/positions": {
            "get": {
                "description": "返回两组标签：self=当前人物在共同条目中担任的职位（含 CV），other=合作人物的职位；同名职位跨条目类型合并，key 可直接回传给 collaboration 的 positions_a/positions_b。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "人物合作"
                ],
                "summary": "人物合作棋盘筛选职位标签",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "人物 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.apiEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.collaborationPositionsData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "无效的 id",
                        "schema": {
                            "$ref": "#/definitions/api.apiEnvelope"
                        }
                    },
                    "404": {
                        "description": "人物不存在",
                        "schema": {
                            "$ref": "#/definitions/api.apiEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.apiEnvelope"
                        }
                    }
                }
            }
        },
        "/api/persons/{id}/collaboration/{other}": {
            "get": {
                "description": "两人物共同参与的条目及双方职务（职位 id 已转常量文本，CV 标记声优出演及角色名）。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "人物合作"
                ],
                "summary": "双人合作",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "人物 A ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "人物 B ID（不能与 A 相同）",
                        "name": "other",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.apiEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.pairCollaborationData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "无效的 id / 两个人物 ID 相同",
                        "schema": {
                            "$ref": "#/definitions/api.apiEnvelope"
                        }
                    },
                    "404": {
                        "description": "人物不存在",
                        "schema": {
                            "$ref": "#/definitions/api.apiEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.apiEnvelope"
                        }
                    }
                }
            }
        },
        "/api/persons/{id}/collaborators": {
            "get": {
                "description": "按共同参与的作品数降序分页返回（含声优等出演关联）。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "人物"
                ],
                "summary": "与「X」合作的人物",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "人物 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 30,
                        "description": "每页数量",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.apiEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.collaboratorsData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "无效的 id",
                        "schema": {
                            "$ref": "#/definitions/api.apiEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.apiEnvelope"
                        }
                    }
                }
            }
        },
        "/api/persons/{id}/roles": {
            "get": {
                "description": "人物参与的全部条目及职务（含 CV 出演，职位 id 已转常量文本）。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "人物合作"
                ],
                "summary": "单人作品",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "人物 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.apiEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.rolesData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "无效的 id",
                        "schema": {
                            "$ref": "#/definitions/api.apiEnvelope"
                        }
                    },
                    "404": {
                        "description": "人物不存在",
                        "schema": {
                            "$ref": "#/definitions/api.apiEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.apiEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.apiEnvelope": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "ok": {
                    "type": "boolean"
                }
            }
        },
...
        "api.characterSearchData": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.characterBrief"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "api.episodeSearchData": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.episodeBrief"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "api.personSearchData": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.personBrief"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AdminToken": {
            "type": "apiKey",
            "name": "X-Admin-Token",
            "in": "header"
        }
    }
}